Exclude from Windows test runs
